requirements.txt
fastapi
uvicorn.run("main:app")

python-multipart
aiofiles